Use map to store resources instead of slice.
This makes the resource retrieval faster since it can be directly
addressed using the file name instead of iterating over the RESOURCE
items.

The other benefit is the use of a static map instead of constructing it
within the init() function. Now we don't need to append each resource to
the RESOURCE variable.

Signed-off-by: Sébastien Gross <seb•ɑƬ•chezwam•ɖɵʈ•org>